+++ codegen/build.sh
@@ -14,5 +14,8 @@ sed -i -r "s/(port_bindings: Option<HashMap<String, )(Vec<InspectHostPort>>>)/\1
 # https://github.com/vv9k/podman-api-rs/issues/113
 sed -i -r "s/(type PortMap = HashMap<String, )String>/\1Option<Vec<PortBinding>>>/g" src/models.rs
 
+# https://github.com/vv9k/podman-api-rs/issues/120
+sed -i -r "s/(ports: Option<HashMap<String, )(Vec<InspectHostPort>>>)/\1Option<\2>/g" src/models.rs
+
 cargo fmt